requirements.txt
Django>=5.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-filter>=24.0
django-cors-headers>=4.0.0
drf-spectacular>=0.27.0
Faker>=20.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
django-extensions>=3.2.0
